move the CameraVisPatch, Shaders, and VisualDiff to the userinterface, to allow mobile devices to take spatial patches too
